+++ src/pages/LiberyPage/tabs.config.tsx
@@ -3,7 +3,7 @@ import OverviewTab from './tabs/Overview'
 import ComponentsTab from './tabs/Components'
 import HooksTab from './tabs/Hooks'
 
-export const LIBRARY_TABS: TabItem[] = [
+export const LiberyTabs: TabItem[] = [
   { key: 'overview', label: 'Overview', content: <OverviewTab /> },
   { key: 'components', label: 'Components', content: <ComponentsTab /> },
   { key: 'hooks', label: 'Hooks', content: <HooksTab /> },
